+++ extensions/cpanel/CHANGELOG.md
@@ -1,3 +1,6 @@
 # cPanel Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- View Domains and their DNS Zones
+- View Email Accounts and their Disk Usage